requirements.txt
boto3==1.28.78
botocore==1.31.78
jmespath==1.0.1
Pillow==10.1.0
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
